% Generated by roxygen2: do not edit by hand
\name{fpkmheatmap}
\alias{fpkmheatmap}
\title{
\packageTitle{countToFPKM}
A FPKM heatmap plot of the highly variable features in RNA-Seq data set
}

\usage{
fpkmheatmap(fpkm_matrix, topvar=30, showfeaturenames=TRUE, return_log = TRUE)
}

\arguments{
\item{fpkm_matrix}{A data matrix normalized by library size and feature length.}
\item{topvar}{Number of highly variable features to show in heatmap plot.}
\item{showfeaturenames}{whether to show the name of features in heatmap plot.
The default value is \code{TRUE}.}
\item{return_log}{whether to use log10 transformation of fpkm.
The default value is \code{TRUE}.}
}

\value{
A FPKM heatmap plot of the highly variable features in RNA-Seq data set.
}

\description{
fpkmheatmap() function returns a heatmap plot of the highly variable features in RNA-Seq data set.
}

\references{
Trapnell,C. et al. (2010) Transcript assembly and quantification by RNA-seq reveals unannotated transcripts and isoform switching during cell differentiation. 
Nat. Biotechnol., 28, 511-515. doi: 10.1038/nbt.1621.
}

\details{
The following function provide the user with a quick and reliable way to generate heatmap plot of the highly variable features in RNA-Seq data set.
This function returns heatmap of FPKM matrix, (by default using Pearson correlation - 1 to calcualte the distance measurments between features, and Spearman correlation -1 for clustering of samples as in fpkmheatmap).
It takes that fpkm numeric matrix from the fpkm() function as input. It then using the var() function to identify the list of highly variable features to plot using Heatmap() function from "ComplexHeatmap" package.
}
